requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.13.5
cassandra-driver==3.29.2
cassio==0.1.10
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.3.0
filelock==3.19.1
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.9.0
gast==0.6.0
geomet==0.2.1.post1
google-pasta==0.2.0
groq==0.31.1
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
keras==3.11.3
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-groq==0.3.8
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.11
langchainhub==0.1.21
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.6
langsmith==0.4.27
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
namex==0.1.0
networkx==3.4.2
numpy==2.0.2
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.3
ormsgpack==1.10.0
packaging==24.2
pillow==11.3.0
propcache==0.3.2
protobuf==5.29.5
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tf_keras==2.18.0
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.1
types-requests==2.32.4.20250809
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
zstandard==0.24.0
